LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-addTransceiver.https-expected.txt


PASS addTransceiver() with string argument as invalid kind should throw TypeError
PASS addTransceiver('audio') should return an audio transceiver
PASS addTransceiver('video') should return a video transceiver
PASS addTransceiver() with direction sendonly should have result transceiver.direction be the same
PASS addTransceiver() with direction inactive should have result transceiver.direction be the same
PASS addTransceiver() with invalid direction should throw TypeError
PASS addTransceiver(track) should have result with sender.track be given track
PASS addTransceiver(track) multiple times should create multiple transceivers
PASS addTransceiver("video") with rid containing invalid non-alphanumeric characters should throw TypeError
PASS addTransceiver("video") with rid longer than 16 characters should throw TypeError
FAIL addTransceiver("video") with duplicate rids should throw TypeError assert_throws_js: function "() =>
      pc.addTransceiver(kind, {
        sendEncodings: [{rid: 'a'}, {rid: 'a'}]
      })" did not throw
PASS addTransceiver("video") with valid rid value should succeed
PASS addTransceiver("video") with multiple rid values should succeed
PASS addTransceiver("video") with valid sendEncodings should succeed
PASS addTransceiver("audio") with rid containing invalid non-alphanumeric characters should throw TypeError
PASS addTransceiver("audio") with rid longer than 16 characters should throw TypeError
FAIL addTransceiver("audio") with duplicate rids should throw TypeError assert_throws_js: function "() =>
      pc.addTransceiver(kind, {
        sendEncodings: [{rid: 'a'}, {rid: 'a'}]
      })" did not throw
PASS addTransceiver("audio") with valid rid value should succeed
PASS addTransceiver("audio") with multiple rid values should succeed
PASS addTransceiver("audio") with valid sendEncodings should succeed
